Test the software and fix most of bugs
We've checked this project in a real machine,and found some bugs.Most of bugs is fixed,but some are not because of the deletion of our hardware.
Anuwau,the robot with it can move normallu now.
All the version are updated now.Next update may not come recently.I'll start a project about RISC-V.

By Harold Zhao